langchain_setup/requirements.txt
langchain>=0.1.20
langchain-google-genai>=0.0.8
google-generativeai>=0.5.4
pypdf>=3.17.0
pandas>=2.2.2
python-dotenv>=1.0.1
tiktoken>=0.5.1
langchain-community>=0.1.20